// mobilekingautoshop.client/src/interfaces/IUser.ts
export default interface IUser {
    Email: string;
    FirstName: string;
    LastName: string;
    PhoneNumber: string;
    Gender: string;
    OtherGender: string;
    Password: string;
    Role: string;
    Address: string;
    City: string;
    State: string;
    ZipCode: string;
    Country: string;
}